# frontend/contract/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_audio_api_v1_voice_upload_post": {
        "properties": {
          "consent_confirmed": {
            "title": "Consent Confirmed",
            "type": "boolean"
          },
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "contentMediaType": "application/octet-stream",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "terms_version": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Version"
          }
        },
        "required": [
          "name",
          "consent_confirmed"
        ],
        "title": "Body_upload_audio_api_v1_voice_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QualityHintOut": {
        "description": "One thing the user can fix in their recording.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "QualityHintOut",
        "type": "object"
      },
      "QualityMeasuresOut": {
        "description": "Physical measures of the uploaded recording.",
        "properties": {
          "clipping_ratio": {
            "title": "Clipping Ratio",
            "type": "number"
          },
          "high_band_db": {
            "title": "High Band Db",
            "type": "number"
          },
          "level_dbfs": {
            "title": "Level Dbfs",
            "type": "number"
          },
          "snr_db": {
            "title": "Snr Db",
            "type": "number"
          },
          "voiced_seconds": {
            "title": "Voiced Seconds",
            "type": "number"
          }
        },
        "required": [
          "voiced_seconds",
          "snr_db",
          "clipping_ratio",
          "high_band_db",
          "level_dbfs"
        ],
        "title": "QualityMeasuresOut",
        "type": "object"
      },
      "QualityOut": {
        "description": "How well the recording suits cloning, with advice (SPEECH_QUALITY_PLAN.md S2.2).",
        "properties": {
          "hints": {
            "items": {
              "$ref": "#/components/schemas/QualityHintOut"
            },
            "title": "Hints",
            "type": "array"
          },
          "measures": {
            "$ref": "#/components/schemas/QualityMeasuresOut"
          },
          "rating": {
            "title": "Rating",
            "type": "string"
          }
        },
        "required": [
          "rating",
          "hints",
          "measures"
        ],
        "title": "QualityOut",
        "type": "object"
      },
...
      "SampleOut": {
        "description": "One clip of the upload and how it compares with the others.",
        "properties": {
          "agreement": {
            "title": "Agreement",
            "type": "number"
          },
          "mismatch": {
            "title": "Mismatch",
            "type": "boolean"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "position",
          "agreement",
          "mismatch"
        ],
        "title": "SampleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VoiceProfileOut": {
        "properties": {
          "consent_confirmed_at": {
            "format": "date-time",
            "title": "Consent Confirmed At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "terms_version": {
            "title": "Terms Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "consent_confirmed_at",
          "terms_version",
          "created_at"
        ],
        "title": "VoiceProfileOut",
        "type": "object"
      },
      "VoiceProfileUploadOut": {
        "description": "The upload response: the profile, plus recording-quality reports when they could be made.\n\n``quality`` is the report of the lowest-rated clip (the one most worth re-recording);\n``samples`` has one entry per uploaded clip, in upload order.",
        "properties": {
          "consent_confirmed_at": {
            "format": "date-time",
            "title": "Consent Confirmed At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "samples": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SampleOut"
            },
            "title": "Samples",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "terms_version": {
            "title": "Terms Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "consent_confirmed_at",
          "terms_version",
          "created_at"
        ],
        "title": "VoiceProfileUploadOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "CloneVoice API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/voice/profiles": {
      "get": {
        "description": "List all active voice profiles for the current user, newest first.",
        "operationId": "get_profiles_api_v1_voice_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VoiceProfileOut"
                  },
                  "title": "Response Get Profiles Api V1 Voice Profiles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Profiles",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/profiles/{profile_id}": {
      "delete": {
        "description": "Soft-delete a voice profile belonging to the current user.",
        "operationId": "delete_profile_api_v1_voice_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Profile",
        "tags": [
          "voice"
        ]
      }
    },
    "/api/v1/voice/upload": {
      "post": {
        "description": "Validate, process and embed one audio sample, or several clips of one speaker.\n\nSend a single clip as ``file`` (as before) and/or several as ``files``. Several clips\nare averaged into one steadier voice embedding (SPEECH_QUALITY_PLAN.md S2.1). If any\nclip fails, none is kept and no profile is created.",
        "operationId": "upload_audio_api_v1_voice_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_api_v1_voice_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceProfileUploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Upload Audio",
        "tags": [
          "voice"
        ]
      }
    },
...
  }
}